Rename last_id_queue.rs to hash_queue.rs
