+++ runtime/src/lib.rs
@@ -2,7 +2,7 @@ mod accounts;
 pub mod appendvec;
 pub mod bank;
 pub mod bloom;
-mod last_id_queue;
+mod hash_queue;
 mod runtime;
 mod status_cache;
